Finish merging evaluator into host
